+++ fiboa_cli/datasets/de_bb.py
@@ -13,8 +13,8 @@ class Converter(AdminConverterMixin, AddHCATMixin, FiboaBaseConverter):
     description = """A Crop Field (German: "Schlaege") is a contiguous agricultural area surrounded by permanent boundaries, which is cultivated with a single crop."""
     license = "DL-DE-BY-2.0"
     provider = "Land Brandenburg <https://geobroker.geobasis-bb.de/gbss.php?MODE=GetProductInformation&PRODUCTID=9e95f21f-4ecf-4682-9a44-e5f7609f6fa0>"
-    ec_mapping_csv = "de.csv"
-    ec_mapping_supplements = ["https://fiboa.org/code/de/de_supplement.csv"]
+    hcat_mapping_csv = "de.csv"
+    hcat_mapping_supplements = ["https://fiboa.org/code/de/de_supplement.csv"]
     # The .cpg claims UTF-8 but the DBF is cp1252 (June 2026 download)
     open_options = dict(encoding="cp1252")