+++ fiboa_cli/datasets/ec_nl_crop.py
@@ -3,7 +3,7 @@ from .nl import NLCropConverter
 
 
 class NLEuroCropConverter(EuroCropsConverterMixin, NLCropConverter):
-    ec_mapping_csv = "nl_2020.csv"
+    hcat_mapping_csv = "nl_2020.csv"
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)